applets/comic: use fallback error placeholder for tiny widgets
Since the original implementation of resizing the widget on showing a
PlaceholderMessage was abandonded, the placeholder could escape the area
of the widget. This looks broken.

Instead, we use a second fallback placeholder that only shows an error
icon if the main one does not fit. Along the way, we get rid of some
fiddling with implicit sizes that is no longer necessary without the
automatic resizing.

BUG: 503478
FIXED-IN: 6.4.0