applets/comic: also show error placeholder in FullView
If the widget is displaying an error message, the full view popup is
still available, but activating it will show an empty popup.

Instead, also show the error Placeholder there. An alternative would be
to disable the popup, but showing the placeholder is useful in case the
widget is small enough to only show the reduced placeholder; this way
the full placeholder contents are still available.