applets/comic: make reason for request explicit
The applet requests data from the engine for several reasons: to display
a particular comic, to check for updates, and to pre-download other
images for a smoother reading experience. These all need to be handled
differently (e.g. only requests for viewing should be immediately
displayed to the user). The requests are currently being disambiguated
by maintaining internal state of the type "Am I checking for updates?"
and "What was the last comic I explicitly requested for viewing?". This
is hard to reason about and has led to severe bugs like infinite
download loops (see 46a85db3).

There is a much simpler solution: we can tell the engine why we're
requesting something, and the engine passes it on to the recipient. This
way, we can avoid internal state and have simpler, more descriptive code
paths.