applets/comic: get rid of mPreviousFailedIdentifier
mPreviousFailedIdentifier is set or cleared after requesting a
particular id for viewing, depending on whether there was an error, but
only read in a handler when the network is reconnected. If another strip
is selected (and available in cache), the variable is cleared. Thus,
this variable is just an unecessarily tracked combination of mCurrent's
isError and id properties.

Instead, get rid of it and just read from mCurrent if necessary.