applets/comic: simplify comic loading
The applet sets the initial comic from the cpp side, but the qml side is
requesting it anyway, so this is unecessary and complicates the control
flow. Similarly, when changing the configuration, the applet half-
switches the current comic, which is again unnecessary because the qml
takes care of everything.

Instead, leave control here completely on the qml side.