applets/comic: provide QML API for stripselector
The "Jump To…" feature currently works completely inside the cpp side
of the applet with a method in the applet class. This makes the flow
of control somewhat less clear, and is bad for tooling as custom methods
are not exported to the type system.

Instead, create a JumpDialog class that is exported to qml, with an API
similar to FileDialog. This makes things clearer on the qml side, and
reduces the number of unqualified lookup messages from tooling a bit.