applets/comic: fix "Jump to" on cached strips
The "Jump to…" feature has a rather elaborate setup to pick an
appropriate dialog for each strip type, so e.g. date-based strips can
be chosen on a calendar. To do this, it needs to know the identifier
type. This breaks on cached strips, which are always set to
StringIdentifier as that's how the cache implementation works. Thus the
correct dialog is only shown for things that were loaded fresh, but not
for ones loaded from cache.

Instead, read the IdentifierType from the plugin metadata again before
passing the data from the engine to the applet if it's a cached strip,
and use that instead of the data from the CachedProvider.