applets/comic: make error placeholder a component
The error placeholder is used twice, both in the fullRepresentation and
in the full view popup. Currently, the actually shared code is rather
small, but as we improve error reporting it will become less so.

This change factors out the shared code into its own component. Along
the way, we fix the full view popup to not show old images (in addition
to the error placeholder) when switching to an image that causes an
error.