applets/comic: show custom connectivity error message
One way that loading an image might fail is if the user is offline or on
otherwise limited connectivity. In those cases, the local cache is still
available so some things can be displayed, but others cannot. The user
can probably figure this out, but a more specific error display seems
still helpful.

This change adds states to ErrorPlaceholder for disconnected and other
non-online states.

Cu-authored by Tobias Fella <tobias.fella@kde.org>