applets/comic: improve error messages for date-based providers
The applet allows the user to select all dates for date-based providers,
and very aggressively tries to access the image for the current day. Not
all these providers publish images on all dates, and certainly not in
sync with each user's local time zone. Some of this (the ability to
manually choose dates that are not available) are likely not generally
fixable, the aggressive attempt to get today's image should be fixable
but would likely require larger changes in the engine.

In the meantime, we can at least improve the error reporting to convey
to users why they are seeing this error message, and inform them that
they could try accessing a different date.