applets/comic: move online change retry handler to qml
The comic applet has a handler such that when the network comes online,
it tries to redownload and display the currently selected comic if the
previous attempt to load it failed. This happens all on the C++ side,
which makes the control flow somewhat hard to reason about as comic
updates can be triggered from both ends, rather than having a single
origin.

This change moves this to a signal handler on the qml side.