applets/comic: make showNextNewStrip a proper function
The "Show next new strip" context menu action calls a C++ signal on the
applet, which only has a signal handler on the qml side. There's no
reason for this to be a signal passing through the C++ applet when all
parts are in qml.

Instead, just have it be a function in the fullRepresentation that is
called directly.