applets/comic: move retry timer to qml
The timer to retry a failed comic download is currently handled on the
C++ side. This makes the flow of control somewhat less clear, and
requires more imperative control - for example, the timer is currently
needlessly running all the time, checking every five minutes whether
there was an error it should retry.

Instead, we can do the same thing from qml and benefit from declarative
bindings to do more (including stopping the timer if there was no error)
with much less code, and have the control of image display happen more
clearly on the qml side.