applets/comic: drop saveConfig function
The only thing the saveConfig function does is save the name of the
currently active comic provider when the configuration dialog is saved.
It's never read. I assume the originally intended feature was to save
the currently selected tab so it could be opened by default on future
Plasma startups, but it at the moment does nothing of the sort. That
would be a somewhat useful feature, but is currently somewhat difficult
to implement due to some complicated dependencies that await some
further infrastructure refactoring.

In the meantime, let's drop this - the existing code (and this function
in general) doesn't do anything and it unlikely to be helpful for
implementing this feature in the future. Also, drop the call to
configChanged when saving the config dialog - Plasma calls this
automaticallly when the saving happens moments later, so reloading the
model unnecessarily happens twice.