applets/comic: move comic methods into comicData
Three methods in the main applet depend only on comicData: saving the
current image to file and launching the website or shop. These fit
semantically better in comicData (as they're actions related to the
currently selected comic) and moving them there has some benefits for
tooling, as they can be registered with that type but are not with the
applet type.

This change moves them there. Along the way, we disable the Open Shop
action in case there is no shop url set for that provider. Along the way
we fix the detection for empty shopUrls to actually work, so the menu
entry is properly disabled if there is no shop set.