applets/comic: don't unnecessarily reload model
The applet currently reloads the model and update checker whenever the
configuration is saved, as some of the changed settings may apply to
them. As model reloads change the current tab, this is rather annoying
when e.g. saving an image (which stores the directory in the
configuration).

Instead, check whether the list of enabled providers and/or update
checking interval were changed, and only reload the parts that are
necessary.