applets/comic: increase the intervals for (re)checking updates
The comic applet is currently very aggressive about checking for updates
and for success on comics that had an error: it checks by default every
30 minutes for updates, and every 5 minutes for the current comic if
there was a loading error (this one cannot be configured differently).
Due to the architecture of the applet, the full comic is downloaded each
time there is no error. This is a lot of downloads for little benefit:
most providers don't update that often, so the same comic is checked and
downloaded again and again, wasting energy and the provider's bandwith,
and I suspect few users care much about having the newest image that
quickly. With the current setup, errors are also rather common for
date-based providers in particular, or cases like APoD where the plugin
fails whenever the daily picture is replaced by a video.

This change increases the default interval between checking everything
for new releases to 3 hours, and the fixed time for retrying a failed
request to 30 minutes. Users who really really care about having the
images as quickly as possible can manually decrease the update
frequency. The update checks on Plasma startup and the retry on re
-establishing internet connection are not affected. As the applet
prior to very recent changes always explicitly wrote all keys to the
appletsrc file, existing installations of the applet will not be
affected by the change to the defaults, only newly added ones. This
change also extends the maximum checking interval to 999 minutes, from
99 before.