applets/mediaframe: reorder and rename config pages
When adding the media frame widget, it does nothing by default except
display a "Configure" button. When the user presses this button, the
resulting configuration dialog does not initially show the most crucial
page for successful use of the widget: actually selecting the media to
show. Instead, they can choose how often they want the images they
haven't selected yet to be swapped, and whether that should happen in a
random or alphabetical order, all irrelevant stuff. The real page that
they need also has an awkward name, "Paths", that is not really
explanatory.

This change swaps the order of the two config pages, so the user is
immediately presented with a large message telling them to add their
media and why they would want to do that. The page is also renamed Media
to be more descriptive. The other page is renamed "Behavior" as General
is not informative and we typically use that as the first entry if we
don't have a better term; for the second page we usually have "Behavior"
(e.g. in Task Manager or Comic). So let's go with that here as well.
And while we're at it, use image-insert-symbolic as the placeholder
icon, as inserting/adding images is exactly what the user needs to do
to use the widget.