applets/mediaframe: use PlaceholderMessage if empty
Mediaframe currently just uses the default configurationNeeded style if
the list of images is empty (but does a custom implementation for some
reason). This doesn't look nice and doesn't tell the user what it is (if
they add several widgets to the desktop at once, for example).

Instead, we copy the same approach the comic widget uses now and use a
PlaceholderMessage with a helpfulAction. As the user can resize it to be
smaller than that placeholder (or get it to be smaller automatically by
dropping it on a desktop already full of widgets), we fall back to the
single button if it does not fit, but with a contextually appropriate
label. The placeholder should just fit into the default-size widget, but
to be on the safe side we bump it up by two gridUnits or so in case the
user has a larger font set.