applets/notes: Fix transparent color in panel and don't offer choice of light/dark
When in a panel, the popup background is always shown, so it doesn't make sense
to offer translucent dark vs. light. Instead, hide the explicit "light" option
and use regular text color.

When in a panel, both "translucent" and "translucent-light" behave as though
"translucent" were selected. This allows roaming the applet between desktop and
panel, i.e. selecting "translucent-light" and then moving it to a panel
where that option wasn't there.