applets/notes: Replace GridDelegate with custom grid
While GridViewKCM is awesome, it doesn't really work for the preview
we're trying to achieve here. We have a bunch of boxes inside boxes
with space reserved for a label underneath that's not there.

Instead, use a custom ItemDelegate with the SVG and placeholder text.
Also set the ItemDelegate text for accessibility purposes.