applets/comic: restore tab on startup and provider changes
Currently, the first provider is always selected on plasmashell startup
and whenever the list of enabled providers is changed. In the latter
case, the applet may end up in a visually broken state where the
visually selected tab and the displayed image do not match. There
already is  a key for saving the currently selected provider, although
it did not work recently and  the remnants of the old method were
removed in 5a9ba90d. With the changes to make the model reload less
frequently, this can now be implemented again on the qml side.

This change adds this functionality back and fixes the glitched provider
list changes.


(cherry picked from commit 1238adda3dba61b8e577ecd4057f76fc9793ada9)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>