applets/fifteenpuzzle: reduce switch size
Fifteen Puzzle has a rather large switch size, so if someone puts it
even on a rather chonky panel, they'll get a huge icon. What the user
intended (if they did it intentionally) was likely to get a smallish
playable puzzle, rather than a giant green question mark icon.

Instead, set the switch sizes to 5 gridUnits, which is what we also use
in similar applets like Comic or Media Frame. It's definitely playable
at around that size. We keep the original values as the default sizes,
so adding it to the desktop should result in the same size as before. In
the future we might want to consider reducing the size of the shuffle
button/timer a bit to make them fit that size even for languages with
rather long words but it should be fine for now - I'd say even a
slightly cut-off version is better than the status quo.

BUG: 360985


(cherry picked from commit c2c8cc2fe058094dcf190f4d063e238b8328e610)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>