applets/fifteenPuzzle: remove fullRep default size
The fullRepresentation sets a prefered size. For panel widgets, this is
largely irrelevant after the first open, as the size is taken from the
last saved value. For desktop widgets, this will resize the widget to
be at least that size on Plasmashell start. Users generally dislike it
when desktop widgets that they may have resized on purpose change their
size in ways they can't predict.

Instead, don't do that and rely on the inital value. If the user thinks
the widget is too small or large for the board size they configured,
they can resize it on their own. Along the way, make a parameter
explicit so Qt doesn't complain about it.


(cherry picked from commit c8a5583d5b4c066448cb832e45a26e1dec4bf94a)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>