applets/comic: fix updates not working
Updates were broken in three ways: the correct default value for the
interval was not set, so auto-updates were off if left at the
default value of 30 minutes; a misspecified conditional led to the
checker not being instantiated, and if auto-updates are off, there is no
way to bypass the cache, so the widget is stuck at the state it was when
auto-updates were turned off. (For some provider types, Jump to still
worked, but not for number-based providers and this is rather
cumbersome).

This change fixes the wrong default values and conditional, and sets
every attempt to read the most recent strip from a comic to bypass the
cache if auto-updates are off.

BUG: 504825