applets/fifteenpuzzle: use radio button for label config
Fifteen Puzzle allows the user to hide the number labels; this makes the
game more challenging (and somewhat more visually appealing) when a
background image is used, but virtually unplayable (an almost unwinnable
guessing game on reasonable board sizes) with a solid background color.
Currently, the settings allow you to go into this essentially broken
state.

Instead, use a radio button to select label with color or hidden label,
and force label display on solid color backgrounds for good measure.
When the user switches from image background to solid color, the label
radio button is automatically switched if necessary, and the invalid
radio button is disabled. And while we're touching this part, do some
routine updates to our current best practices, like setting translation
contexts and setting text and display properties on icon-only buttons.

BUG: 504916
FIXED-IN: 6.5.0