applets/fifteenpuzzle: hide board instead of covering it when solved
Fifteen Puzzle places a Rectangle over the board when the game is solved
to hide the pieces and display the victory message (and full image if
applicable). This doesn't look nice when put on the desktop with no
image, as the rectangle uses the Theme background color, rather than
the slightly transparent desktop widget color. It also leads to janky
resizing behavior when solved, as the covered pieces may peek out from
behind the rectangle as they are dynamically resized.

Instead, just hide the whole board when it is solved, and replace it
with a victory sceen that does not force a background.

BUG: 504921