applets/fifteenPuzzle: set minimum size for fullRepresentation
The fullRepresentation does not have a minimum size. This is not a
problem in the case of the desktop widget, which will fall back to the
compactRepresentation, but it makes no sense for the popup.

Instead, we set it to have a minimum size equal to the switch size.
Along the way we get rid of some not particularly useful debug messages.

BUG: 504926