applets/fifteenpuzzle: make tiny sizes work better
We allow users to make the applet really small, which can leave it
unplayable in some configurations. Part of that is a clear "Don't do
that then" case (if you want to play on a 18×18 grid, don't make it
barely larger than a stamp) but we can do a much better job at making it
fit on the smallest sizes.

This change enables word wrap for the victory screen labels, replaces
the shuffle button with an icon-only one if its row would not fit, and
fixes a spurious binding (my bad) that can cause the field to escape the
plasmoid size if you resize it in some ways. Along the way, fix an
implicit parameter warning.

BUG: 504920


(cherry picked from commit d0ad03e9f7f678270fcb856d13ae77785a1bee14)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>