applets/webbrowser: use symbolic icon
Web Browser defaults to using the favicon for the compactRepresentation,
but the fallback, and default for a custom icon, is the full-color
internet icon. This does not fit in with out usually symbolic icons.

Instead, use the symbolic browser icon.


(cherry picked from commit ab71831058b87eede7afa41271d9745eb6b1ef49)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>