plasmacalendarplugins: make plugin descriptions more informative
The calendar plugin descriptions are currently not exposed to users, but
we're preparing to do this to make the functionality of the plugins
clearer. The current descriptions have some unnecessary bits (that they
are plugins) and could instead be a bit more explicit about what exactly
they do when enabled.

This change updates the plugin descriptions.


Co-authored-by: Nate Graham <nate@kde.org>