applets/comic: fix applet on first start
Currently the applet pretty much does not work when first added, only
after restarting Plasma. This is for a major and a minor reason. The
major one is that when adding from widget explorer, the initialization
of qml and the C++ applet happens with slightly different timing; the
model is initialized in init(); when first added this is still null as
init() hasn't happened yet and it will never be updated as it is a
CONSTANT property. The minor reason only applies to cases where a single
comic is added, here our timing for setting the (invisible) tab is a bit
off.

Instead, we jiggle things around so we can initialize the model in the
member initializer list (this works now that we can set the providers
later), and adjust the timing of loading a single tab to when it
definitely works.