applets/fifteenPuzzle: fix loading errors
Fifteen Puzzle causes a lot of errors to be printed on Plasma startup
when a background image is set. This is for two reasons: one, it tries
to get negative size images while the Layouting is still trying to
figure out the right piece sizes, which confuses the backend, and two,
it tries to also get the pixmap for the missing image, which due to a
numbering mismatch between the backend and frontend works out to
a negative id, which the backend can't handle (and is pointless in the
first place, as neither the missing piece nor its background will ever
be shown, or it wouldn't be the missing piece).

Instead, we just don't request the image if height, width, or piece
number would be zero or less. This displays the pieces just as before,
but with no spurious error messages (and possibly marginally more
efficiently).