applets/fifteenPuzzle: fix image update and allow '-' in image name
Manual cherry-pick of MR !819

FIRST PROBLEM
In case the image had '-' character in it, it would silently fail to
load the image ("falling back" to solid color). This is due of the use
of split over the arguments passed to image loader. This is solved by
using a maxSplit as the path argument passed to the cpp image loader is
in the last position.

SECOND PROBLEM
NOTE: if you are not capable to solve, just reduce the size of the grid
in the settings to 2, then solving it stupidly easy. It is even possible
that the puzzle generated is solved. If it is the case then just move
and unmove a piece to solve.

Change authored by Matteo Di Noia <dmatteo002@gmail.com>