applets/notes: fix fullRepresentation in panel
The fullRepresentation does not take into account that it could be
placed in a very thick panel. It sets minimum width in that case (to
constrain the popup), but they also apply to the fullRepresentation in
the panel directly, is much smaller than that case can handle, and
causes it to go on a wild resizing spree that locks up plasmashell
completely.

Instead, we do not set minimum widths in that case, which does the only
sensible thing: a square widget that fits the panel size. While we're at
it, also tweak the other panel checks that only apply in case of
compactRepresentation-in-panel; there is for example no point to having
the pin button for fullRepresentation-in-panel as it's not going to
close anyway.

BUG: 505132
FIXED-IN: 6.4.0