applets/fifteenpuzzle: disable animation when not needed
Disable animation during startup and during resize of the widget.

This is done as animation during boot/reload cause lag spike and the animation is "broken" causing all tiles to scale from the origin point.
Animating during resize made it possible for the pieces to exit the widget

Before on load (black background is because i need to restart plasmashell to show this):

![bad-start-1](/uploads/2cadd5d2afbadebc801a421c2ecb9297/bad-start-1.png)
![bad-start-2](/uploads/a7dbec4d3e6eeaa6984057334c7057c5/bad-start-2.png)

Before on resize:

![bad-resize-1](/uploads/f4c1baa872b68202b13ffe17994a3b9c/bad-resize-1.png)
![bad-resize-2](/uploads/10933bb585a08e0d22bdc924e4bbbea9/bad-resize-2.png)

Now both behave as expected
RESIZE 

![good-resize](/uploads/ad58f1827607a76cf5ef85813acbaeef/good-resize.png)

START (black background is because i need to restart plasmashell to show this)

![good-start](/uploads/fe185303b6f3630130f6f102bac4bb38/good-start.png)


This is done by adding a timer that disable animation for some time after the start of a resize of a piece (which also happen on creation), this doesn't stop any other animation in any other time.