applets/fifteenPuzzle: don't load image if disabled
If the user has an image filename set, but the image disabled in
settings, the full image for the completed board is loaded anyway and
just made invisible. This is a bit wasteful (and can cause spurious
runtime warnings if the file name does not exist anymore). The
individual pieces are not affected, as they are controlled through a
loader that's only active when the setting is active.

Instead, don't set the image source if we're not planning to display the
image. We also set empty string instead of null in that case, resolving
another warning (as QUrls can't be null).