applets/weather: remove icon from missing forecast placeholder
If the forecast is unavailable (e.g. due to flakey provider), the widget
will show a placeholder informing the user. This contains a large
weather-none-available icon. As a consequence, the forecast is much
larger than usual, which will increase the implicit and thus minimum
size, potentially rearranging people's desktop widget arrangement.

Completely getting the widget to not resize itself for desktop is likely
a larger undertaking. But we can reduce the impact of this particular
case by omitting the large icon. If the provider is fully not reporting
data, there will already be a (non-symbolic) weather-none-available icon
visible from the current weather view, and the double similar icons are
somewhat confusing anyway. The placeholder text and explanation by
themselves should in many cases be somewhat similar size to the forecast
so even if it resizes a little it's less likely to substantially bump
other widgets out of the way.

CCBUG: 505332