applets/webbrowser: don't use fullRepresentation in panel
Web Browser's fullRepresentation in a panel hard-freezes Plasma in an
infinite representation switching loop, similar to how the notes applet
did before (bug 505132). Unfortunately, the same solution does not seem
to work here, it freezes even with minimum layout sizes unset, and none
of my attempts at finding a proper solution went anywhere.

Instead, force the compactRepresentation when used in a panel. This is
clearly not the optimal solution, but better than shipping a hard-freeze
trap that may leave users unable to use their computer if they try to
get a web page as a sidebar (or just play around a bit with different
settings).