applet/fifteenPuzzle: don't freeze Plasma in large panel
When the fullRepresentation is used directly in a very wide panel,
especially if other widgets are also present, Plasma may freeze as it
gets stuck on switching between the representations - the
compactRepresentation is large enough that it wants to switch, but the
fullRepresentation ends up too small and switches back, forever.

Unsetting Layout minimum sizes in those cases prevents the issue from
happening.