applets/dict: Don't put fullRepresentation in panel
The panel can't easily get keyboard focus, so you can't actually click
and type into the search field; you have to use a separate shortcut to
focus the panel and tab to the applet (or arrow key if nothing else
steals arrow keys). This makes it much worse than just showing the
compactRepresentation where you can click-and-type, or open it directly
with a keyboard shortcut, rather than having to use an undiscoverable
and quite cumbersome method.

So always use the compactRepresentation in a panel.