applets/calc: don't freeze Plasma & don't auto-resize desktop widget
When the fullRepresentation is used directly in a very wide panel,
especially if other widgets are also present, Plasma may freeze as it
gets stuck on switching between the representations - the
compactRepresentation is large enough that it wants to switch, but the
fullRepresentation ends up too small and switches back, forever.

Unsetting Layout minimum sizes in those cases prevents the issue from
happening. We get rid of the preferred sizes completely, in favor of
setting a default size on the PlasmoidItem - this way we still get the
same initial size, but the content will not auto-resize on plasmashell
load and potentially disrupt users desktop widget arrangements.


(cherry picked from commit b0841c8e323d681cff290d03aae28a66ce6cd513)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>