Port weather applet from dataengine
## General information
This port uses ions as a Qt Plugins instead of dataengines. Each Ion is running in another thread so there is no need to worry 
about blocking operations. If Ion is not in use for now it's thread is stopped. If Ion is not needed (no one uses it) it is removed
and plugin is unloaded.

The main logic of Ions (forecast request and parsing) are not changed, but there are some changes to adapt them to a new architecture.

An additional library weatherdata was created to share data between qml and Ion. The main classes are `Locations` and `Forecast` which
represent locations and forecasts respectively. `Locations` is just a `QAbstractListModel` where `Forecast` is a `Q_GADGET` which are
returned by an Ion. 

The request and update of forecasts and locations is happening by request to `WeatherDataMonitor` - the main class which loads, unloads 
plugins, controls theirs live time, create IonControl's etc. It is a singleton.

The request and update is happening through `LocationsData` and `ForecastData` classes. They are like containers with locations and forecast 
respectively. When locations or forecast is updated they emit signal. 

`IonControl` is a class that manages an ion. It handles multiple requests and controls ion's thread.

`IonControl` requests forecasts and locations from Ion through QPromise and then update ForecastData and LocationsData with them in the 
main thread so no need to use mutexes in `LocationsData` and `ForecastData`.

`LocationsControl` is just a QConcatenateTablesProxyModel which manages `LocationsData` for different providers (ions) and update its model
with models from them.

`ForecastControl` manages `Forecast` for an applet instance.

This MR also contains another MR: https://invent.kde.org/plasma/kdeplasma-addons/-/merge_requests/761

I have tested it and it works for me but it would be good if someone else test this.

I have added some high-level diagrams to show how the weather applet works with this MR below. 

## The structure and relationships among classes involved in this port
![Editor___Mermaid_Chart-2025-06-16-184539](/uploads/c7d35acfc34316d042d0417c03003b91/Editor___Mermaid_Chart-2025-06-16-184539.png)

## The forecast request diagram 
![Editor___Mermaid_Chart-2025-06-16-193004](/uploads/03763b9493721009d27f4b786daf8727/Editor___Mermaid_Chart-2025-06-16-193004.png)

## The locations request diagram
![Editor___Mermaid_Chart-2025-06-16-194319](/uploads/20dec5d75ae9620948903d38010ee8dd/Editor___Mermaid_Chart-2025-06-16-194319.png)

## A screenshot with running a series of weather widgets for various global locations.
![weatherapplet](/uploads/f99cfd4064f96a052d985f6a98c1d963/weatherapplet.png){width=1920 height=1030}

## Closes
https://invent.kde.org/plasma/plasma-workspace/-/issues/109