applets/comic: fix view request backlog with tabs
Amends 832de4ea1155375e1e5cba7785ba6cac9d61cb36

The comic applet retries requests with a short delay if a request for
the same id is active, as can happen if the user switches the image
while an update check or prefetch is running. This causes an issue if
the active tab was changed in the mean time, as it is not checked
whether a preview request is now stale, and get the applet in a confused
state where it tries to prefetch or even display comic (strip) ids from
a different comic. This happens in particular during plasmashell startup
if the last tab the user was on is not the first one.

Instead, we explicitly check whether a view request actually applies to
the currently selected comic, and always prefetch for the comic that a
strip id is actually for, just to be on the safe side. This fixes the
issue.