applets/mediaframe: center delegate labels on media config page
The media config page has listview delegates (SwipeListItem) with the
name of the file/folder and a remove button. It overrides the
contentItem, and does not explicitly center the label, so it's somewhat
awkwardly placed at the top of the delegate compared to the much larger
remove button.

Instead, center the label in the delegate.