applets/grouping: fix adding applets
Amends 1d6a930dcaa94043736d856f3f35481fb7e23ec9

When this applet was converted to plasma_add_applet, the file
PlasmoidItem.qml was moved from the items subfolder directly into
the qml folder. The file is looked up by name for dynamic creation,
and its path in the source code was not adjusted for this move,
breaking the applet.

This change fixes the path.

BUG: 513172
FIXED-IN: 6.5.90