applets/weather: check for a night icon before return an empty icon
Sometimes a provider does not return a condition icon for current conditions.
In that case, the icon from the forecast of the first day is used. But it only
checks for the icon from the daytime forecast which can lead to an empty icon
because a daytime forecast is not always present. In that case, also check the
icon from the night forecast before returning an empty icon.

BUG: 513069