kwin/windowswitchers: allow big_icons to use multiple rows
When there is not enough space to show all icons in the big_icons
switcher, it cuts off the list and scrolls through it as the user
moves the currentIndex. This makes it hard to see all the open windows
even if there is in principle enough space on the screen to show them
all.

Instead, we use a gridView, similar to how the thumbnail grid works. In
contrast to thumbnail grid, which tries to build the best grid, we still
want this to feel like a list, so we do not copy its layouting algorithm
but instead just fill up a row of icons as usual, only it is allowed to
wrap onto new rows as necessary. The tabbox does not handle dialog
height resizing well, so we wrap it in an instantiator that creates it
fresh, again similarly to how the thumbnail grid handles it.

BUG: 513436