applets/weather: Fix visibility for zero degrees temperatures
When showing temperatures in the forecast delegates, we cannot
check for "trueness" of the value in order to hide it, as 0ºC
(or even 0ºF) are valid temperatures.

Check instead if the value has been set. Javascript's `isNaN()`
is very handy, since it works for both `nan`vand `undefined`.

BUG: 512871
FIXED-IN: 6.5.90