kwin/windowswitchers: handle null model in compact
Compact switcher uses TextMetrics for sizing, picking the longest entry
from the model using the model method. This will cause a runtime error
in case the model no longer exists, but the TextMetrics still does.

Instead, use optional chaining and only access the model if it exists.
We also do the same for the handlers for consistency.